forex-smc-nn/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
joblib>=1.3.0
yfinance>=0.2.40
